fix bug in new code generation
Actually, the bug is in the ifft used by new new code generator, which
assumed that its input was not reused.  Actually, it assumed that
calling the "free" method on its input would ensure that the code
would crash if it tried to reuse the input.  This didn't work, and we
*were* reusing the input.  I now just use a scratch array, so we don't
trash our input.

Also added a new test, which now ensures out functional agrees with
the old code generator for homogeneous inputs at least.
Tested on darcy. Took 0:25.63.