extend fuzzy-fmt radial-distribution.py to plot with new data
This causes it to use the new-radial-wca.cpp data and MC data from FCC
starting positions.  Sadly, I accidentally added a bunch of incomplete
data a few commits back.  This commit does not add the correct data
(which is not yet available) but does change the build to use the
incomplete data that was already completed.

Tested on bennet. Took 0:24.98.